requirements.txt
django>=5.0.0
torch>=2.1.0
torchvision>=0.16.0
Pillow>=10.0.0
python-dotenv>=1.0.1

